+++ plugins/GymFactory/include/gympp/Metadata.h
@@ -105,8 +105,8 @@ private:
     std::string modelFileName;
     std::string worldFileName;
 
-    size_t gazeboUpdateRate;
-    size_t environmentUpdateRate;
+    double gazeboUpdateRate;
+    double environmentUpdateRate;
 
     SpaceMetadata actionSpace;
     SpaceMetadata observationSpace;
@@ -117,8 +117,8 @@ public:
     inline std::string getClassName() const { return className; }
     inline std::string getModelFileName() const { return modelFileName; }
     inline std::string getWorldFileName() const { return worldFileName; }
-    inline size_t getGazeboUpdateRate() const { return gazeboUpdateRate; }
-    inline size_t getEnvironmentUpdateRate() const { return environmentUpdateRate; }
+    inline double getGazeboUpdateRate() const { return gazeboUpdateRate; }
+    inline double getEnvironmentUpdateRate() const { return environmentUpdateRate; }
     inline SpaceMetadata getActionSpaceMetadata() const { return actionSpace; }
     inline SpaceMetadata getObservationSpaceMetadata() const { return observationSpace; }
 
@@ -148,12 +148,12 @@ public:
         this->worldFileName = worldFileName;
     }
 
-    inline void setGazeboUpdateRate(const size_t gazeboUpdateRate)
+    inline void setGazeboUpdateRate(const double gazeboUpdateRate)
     {
         this->gazeboUpdateRate = gazeboUpdateRate;
     }
 
-    inline void setEnvironmentUpdateRate(const size_t environmentUpdateRate)
+    inline void setEnvironmentUpdateRate(const double environmentUpdateRate)
     {
         this->environmentUpdateRate = environmentUpdateRate;
     }
@@ -168,8 +168,8 @@ public:
         ok = ok && !worldFileName.empty();
         ok = ok && actionSpace.isValid();
         ok = ok && observationSpace.isValid();
-        ok = ok && (gazeboUpdateRate != 0);
-        ok = ok && (environmentUpdateRate != 0);
+        ok = ok && (gazeboUpdateRate != 0.0);
+        ok = ok && (environmentUpdateRate != 0.0);
         return ok;
     }
 };
